Move VVSG2 auth flows feature flag to libs/ui
