+++ frontends/election-manager/src/env.d.ts
@@ -4,7 +4,6 @@ declare namespace NodeJS {
     readonly REACT_APP_VX_CODE_VERSION?: string;
     readonly REACT_APP_VX_CONVERTER?: string;
     readonly REACT_APP_VX_DEV?: string;
-    readonly REACT_APP_VX_ENABLE_VVSG2_AUTH_FLOWS?: string;
     readonly REACT_APP_VX_ENABLE_WRITE_IN_ADJUDICATION?: string;
     readonly REACT_APP_VX_MACHINE_ID?: string;
   }
