+++ libs/ui/src/env.d.ts
@@ -3,5 +3,6 @@ declare namespace NodeJS {
     readonly NODE_ENV: 'development' | 'production' | 'test';
     readonly REACT_APP_VX_DEV?: string;
     readonly REACT_APP_VX_DISABLE_CARD_READER_CHECK?: string;
+    readonly REACT_APP_VX_ENABLE_VVSG2_AUTH_FLOWS?: string;
   }
 }
