+++ libs/ui/src/index.ts
@@ -51,3 +51,4 @@ export * from './usbcontroller_button';
 export * from './remove_card_screen';
 export * from './invalid_card_screen';
 export * from './unlock_machine_screen';
+export { areVvsg2AuthFlowsEnabled } from './config/features';
